+++ src/Fader.h
@@ -15,7 +15,7 @@ public:
 private:
     bool IsFading;                      // Gibt an, ob gefadet wird
     decimal From, To, Value;            // Intervall in dem gefadet wird [From;To] und aktueller Wert
-    unsigned short FadeTime;            // Zeit in ms, die für den Fade benötigt wird
+    unsigned long FadeTime;            // Zeit in ms, die für den Fade benötigt wird
     unsigned long FadeStartTime;        // Zeit in ms, zu der der Fade gestartet wurde
 
     // Konstruktor
@@ -28,10 +28,10 @@ public:
     void Update();
 
     // Startet den Fade vom aktuellen Wert zum neuen Ziel
-    void Start(unsigned short fadeTime, decimal to);
+    void Start(unsigned long fadeTime, decimal to);
 
     // Startet den Fade im angegebenen Intervall
-    void Start(unsigned short fadeTime, decimal from, decimal to);
+    void Start(unsigned long fadeTime, decimal from, decimal to);
 
     // Stoppt den Fade
     void Stop();
